refactor(frontend): move Shabad OS data constants into separate file
