+++ app/frontend/src/Overlay/Line.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { string, bool, objectOf, func } from 'prop-types'
 import classNames from 'classnames'
 
-import { LANGUAGE_NAMES, TRANSLATION_ORDER, TRANSLITERATION_ORDER } from '../lib/consts'
+import { LANGUAGE_NAMES, TRANSLATION_ORDER, TRANSLITERATION_ORDER } from '../lib/data'
 import { partitionLine, classifyWords } from '../lib/line'
 
 import './Line.css'
