+++ app/frontend/src/Overlay/index.js
@@ -6,7 +6,7 @@ import Line from './Line'
 import ThemeLoader from './ThemeLoader'
 
 import { SettingsContext, StatusContext } from '../lib/contexts'
-import { LANGUAGES } from '../lib/consts'
+import { LANGUAGES } from '../lib/data'
 import { useTranslations, useCurrentLine } from '../lib/hooks'
 import { customiseLine, getTransliterators } from '../lib/line'
